+++ packages/backend/src/clients/Slack/SlackClient.ts
@@ -11,6 +11,12 @@ type SlackClientDependencies = {
     lightdashConfig: LightdashConfig;
 };
 
+const CACHE_TIME = 1000 * 10 * 60; // 10 minutes
+const cachedChannels: Record<
+    string,
+    { lastCached: Date; channels: SlackChannel[] }
+> = {};
+
 export class SlackClient {
     slackAuthenticationModel: SlackAuthenticationModel;
 
@@ -57,6 +63,17 @@ export class SlackClient {
     }
 
     async getChannels(organizationUuid: string): Promise<SlackChannel[]> {
+        if (
+            cachedChannels[organizationUuid] &&
+            new Date().getTime() -
+                cachedChannels[organizationUuid].lastCached.getTime() <
+                CACHE_TIME
+        ) {
+            return cachedChannels[organizationUuid].channels;
+        }
+
+        Logger.debug('Fetching channels from Slack API');
+
         if (this.slackApp === undefined) {
             throw new Error('Slack app is not configured');
         }
@@ -66,7 +83,7 @@ export class SlackClient {
                 organizationUuid,
             );
 
-        const channels = await this.slackApp.client.conversations.list({
+        const conversations = await this.slackApp.client.conversations.list({
             token: installation?.token,
             types: 'public_channel',
             limit: 500,
@@ -75,12 +92,26 @@ export class SlackClient {
         const users = await this.slackApp.client.users.list({
             token: installation?.token,
         });
-        return [...(channels.channels || []), ...(users.members || [])].reduce<
-            SlackChannel[]
-        >(
-            (acc, { id, name }) => (id && name ? [...acc, { id, name }] : acc),
-            [],
-        );
+
+        const sortedChannels = (conversations.channels || [])
+            .reduce<SlackChannel[]>(
+                (acc, { id, name }) =>
+                    id && name ? [...acc, { id, name: `#${name}` }] : acc,
+                [],
+            )
+            .sort((a, b) => a.name.localeCompare(b.name));
+
+        const sortedUsers = (users.members || [])
+            .reduce<SlackChannel[]>(
+                (acc, { id, name }) =>
+                    id && name ? [...acc, { id, name: `@${name}` }] : acc,
+                [],
+            )
+            .sort((a, b) => a.name.localeCompare(b.name));
+
+        const channels = [...sortedChannels, ...sortedUsers];
+        cachedChannels[organizationUuid] = { lastCached: new Date(), channels };
+        return channels;
     }
 
     async joinChannels(organizationUuid: string, channels: string[]) {