+++ packages/frontend/src/components/SimpleStatistic/BigNumberContextMenu.tsx
@@ -24,6 +24,8 @@ export const BigNumberContextMenu: FC<BigNumberContextMenuProps> = ({
     const { resultsData, bigNumberConfig } = useVisualizationContext();
     const { openUnderlyingDataModel, tableName } = useMetricQueryDataContext();
     const { data: explore } = useExplore(tableName);
+    const { user } = useApp();
+    const { projectUuid } = useParams<{ projectUuid: string }>();
 
     const { track } = useTracking();
     const { user } = useApp();
@@ -99,6 +101,11 @@ export const BigNumberContextMenu: FC<BigNumberContextMenuProps> = ({
                     <DrillDownMenuItem
                         row={resultsData?.rows[0]}
                         selectedItem={selectedItem}
+                        trackingData={{
+                            organizationId: user?.data?.organizationUuid,
+                            userId: user?.data?.userUuid,
+                            projectId: projectUuid,
+                        }}
                     />
                 </Menu>
             }